+++ test/extensions/filters/network/tcp_proxy/config_test.cc
@@ -12,58 +12,6 @@ namespace Extensions {
 namespace NetworkFilters {
 namespace TcpProxy {
 
-class RouteIpListConfigTest : public ::testing::TestWithParam<std::string> {};
-
-INSTANTIATE_TEST_CASE_P(IpList, RouteIpListConfigTest,
-                        ::testing::Values(R"EOF("destination_ip_list": [
-                                                  "192.168.1.1/32",
-                                                  "192.168.1.0/24"
-                                                ],
-                                                "source_ip_list": [
-                                                  "192.168.0.0/16",
-                                                  "192.0.0.0/8",
-                                                  "127.0.0.0/8"
-                                                ],)EOF",
-                                          R"EOF("destination_ip_list": [
-                                                  "2001:abcd::/64",
-                                                  "2002:ffff::/32"
-                                                ],
-                                                "source_ip_list": [
-                                                  "ffee::/128",
-                                                  "2001::abcd/64",
-                                                  "1234::5678/128"
-                                                ],)EOF"));
-
-TEST_P(RouteIpListConfigTest, TcpProxy) {
-  std::string json_string = R"EOF(
-  {
-    "stat_prefix": "my_stat_prefix",
-    "route_config": {
-      "routes": [
-        {)EOF" + GetParam() +
-                            R"EOF("destination_ports": "1-1024,2048-4096,12345",
-          "cluster": "fake_cluster"
-        },
-        {
-          "source_ports": "23457,23459",
-          "cluster": "fake_cluster2"
-        }
-      ]
-    }
-  }
-  )EOF";
-
-  Json::ObjectSharedPtr json_config = Json::Factory::loadFromString(json_string);
-  NiceMock<Server::Configuration::MockFactoryContext> context;
-  ConfigFactory factory;
-  Network::FilterFactoryCb cb = factory.createFilterFactory(*json_config, context);
-  Network::MockConnection connection;
-  EXPECT_CALL(connection, addReadFilter(_));
-  cb(connection);
-
-  factory.createFilterFactory(*json_config, context);
-}
-
 TEST(ConfigTest, ValidateFail) {
   NiceMock<Server::Configuration::MockFactoryContext> context;
   EXPECT_THROW(ConfigFactory().createFilterFactoryFromProto(